requirements.txt
PIL
numpy 
matplotlib.pyplot 
mpl_toolkits.mplot3d 
cv2